requirements.txt
openai>=2.0.0
pyyaml>=6.0.3
